test: split chat hr analytics smoke scenario
